+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Rename `IOMessage` widget to `I18nInput`, and add a prop to replace `BaseInput` with `TextArea`.
 
+### Fixed
+
+- Prevent invalid values at `TextArea``.
+
 ## [3.13.7] - 2019-06-06
 
 ### Fixed