+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 
-- `TextArea`s horizontal resizing.
+- `TextArea`'s horizontal resizing.
 
 ## [3.13.7] - 2019-06-06