+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.14.0] - 2019-06-06
+
 ### Added
 
 - `RichText` widget, which, for now, is just a `TextArea` with i18n.